Renamed Logger class to FileLogger, added WindowLogger
The new window logger class allow the creation of a form with a
richtextbox insde, which can then pick up exe output or be printed to
directly. The form itself cannot be accessed directly, but the size and
positioning can be changed programmatically. The window logger can also
be bound to a checkbox, which can then be used to toggle the log window
(it is updated whenever the log is closed or opened).